pyproject.toml
[build-system]
requires = [
"setuptools>=80.9.0,<81",
"cython>=3.1.2,<4",
"numpy>=2.2.6,<3"]
build-backend = "setuptools.build_meta"


[project]
name = "quapopt"
authors = [{ name = "Filip B. Maciejewski", email = "fmaciejewski@usra.edu" }]
description = "Various tools for running quantum approximate optimization experiments and simulations."
readme = "README.md"
version = "0.2.1"
requires-python = "==3.12"


dependencies = [
    "cython >=3.1.2,<4",
    "jupyter>=1.1.1,<2",
    "notebook>=7.4.4,<8",
    "numpy>=2.2.6,<3",
    "pandas>=2.3.0,<3",
    "pydantic>=2.11.7,<3",
    "tqdm>=4.67.1,<5",
    "networkit>=11.1,<12",
    "rustworkx>=0.16.0,<0.17",
    "sympy>=1.14.0,<2",
    "qutip>=5.2.0,<6",
    "matplotlib>=3.10.3,<4",
    "pylatexenc>=2.10,<3",
    "plotly>=6.2.0,<7",
    "optuna>=3.6.1,<4",
    "python-dotenv>=1.1.1,<2",
    "qiskit>=2.1.0,<3",
    "qiskit-ibm-runtime>=0.40.1,<0.41",
    "kaleido>=1.0.0,<2",
    "pytest>=8.4.1,<9",
    "nbstripout>=0.8.1,<0.9",

    #The following packages are downgraded compared to conda channels
    #as they're not available in pypi as of now
    "scipy>=1.15,<2",
    "networkx>=3.4.2,<4",

    #Custom version of mqlib
    "mqlib @ git+https://github.com/fbm2718-usra/pymqlib-quapopt.git"
    #"mqlib@file:./quapopt/additional_packages/pymqlib"
]


[project.optional-dependencies]
mlvl = [
    "pygad>=3.4.0,<4",
    "scikit-learn>=1.7.0,<2",
    "sortedcontainers>=2.4.0,<3"
]


gpu = [
    "cupy-cuda12x",
    "numba-cuda>=0.16.0,<0.17",
]

cpu = [

    #"torch>=2.7.1,<3"

]

visualization = ['cmocean']

qiskit_qopt = [
    "python-sat>=1.8.dev17,<2",
    "qopt-best-practices @ git+https://github.com/qiskit-community/qopt-best-practices.git"]

qiskit_aer = ["qiskit-aer>=0.17.1,<0.18"]

#This does not work currently
qiskit_aer_gpu = ["qiskit-aer>=0.17.1,<0.18"]


full = [
    "quapopt[cpu]",
    "quapopt[mlvl]",
    "quapopt[qiskit_qopt]",
    "quapopt[qiskit_aer]",
    "quapopt[visualization]"
]
full_gpu = [
    "quapopt[mlvl]",
    "quapopt[qiskit_qopt]",
    "quapopt[qiskit_aer]",
    "quapopt[gpu]",
   # "quapopt[qiskit_aer_gpu]"
        "quapopt[visualization]"

]



